+++ tests/test_context.py
@@ -5,13 +5,13 @@ from quickie.context import Context
 class TestContext:
     def test_copy(self):
         context = Context(
-            cwd="test",
+            wd="test",
             env={"MY_VAR": "value"},
             inherit_env=False,
         )
         context_copy = context.copy()
         assert context is not context_copy
-        assert context.cwd == context_copy.cwd
+        assert context.wd == context_copy.wd
         assert context.env == context_copy.env
         assert context.env is not context_copy.env
         assert isinstance(context.env, ChainMap)